+++ package.json
@@ -13,7 +13,6 @@
     "bem-cn-lite": "^4.1.0",
     "next": "^12.3.4",
     "next-global-css": "^1.3.1",
-    "next-images": "^1.8.4",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-timeago": "^7.1.0",
@@ -34,12 +33,12 @@
     "dotenv": "^16.0.3",
     "eslint": "^8.27.0",
     "eslint-plugin-testing-library": "^5.9.1",
+    "file-loader": "^6.2.0",
     "gh-pages": "^4.0.0",
     "nano-staged": "^0.8.0",
     "next-compose-plugins": "^2.2.1",
     "next-transpile-modules": "^10.0.0",
     "prettier": "^2.7.1",
-    "react-snap": "^1.23.0",
     "sass": "^1.56.1",
     "sass-loader": "^13.2.0",
     "stylelint": "^14.15.0",
@@ -53,14 +52,17 @@
   },
   "scripts": {
     "start": "next dev",
-    "build": "GENERATE_SOURCEMAP=false next build",
-    "export": "next export",
+    "build": "next build",
+    "export": "next export -o build/",
     "fetch-data": "node ./scripts/prefetch-data.mjs",
+    "prestart": "npm run fetch-data",
+    "prebuild": "npm run fetch-data",
+    "pretest": "npm run fetch-data",
     "lint:js": "eslint --ext .js,.jsx,.ts,.tsx .",
     "lint:styles": "stylelint '{styles,src}/**/*.scss'",
     "lint:prettier": "prettier --check '**/*.md'",
     "lint": "npm run lint:js && npm run lint:styles && npm run lint:prettier",
-    "predeploy": "npm run build",
+    "predeploy": "npm run build && npm run export",
     "deploy": "echo 'gravity-ui.com' > ./build/CNAME && gh-pages -d build"
   },
   "browserslist": {
